Add advanced tier usage option
Using the SSN parameter advanced tier is required when the key exceeds
4096 characters which is the maximum size of standard tier SSN
parameters. This adds the CanUseAdvancedTier parameter to the
PersistOptions. When creating the parameter, the length of the key
will be evaluated and the advanced tier will be used if required and
configured. When the length exceeds the standard tier size and advanced
tier usage is not configured, an exception will be thrown. When the key
size exceeds the advanced tier, an exception will be thrown aswell.